Login

+++ taudinpurkauspeli/backend/controllers/cases.js
@@ -1,15 +1,19 @@
-const jwt = require('jsonwebtoken')
+/* eslint-disable consistent-return */
 const caseRouter = require('express').Router();
 const db = require('../models');
-const config = require('../utils/config')
+const helper = require('../utils/token');
 
 const Case = db.cases;
-const User = db.users;
 const { Op } = db.Sequelize;
 
 // Save a new case
 caseRouter.post('/', (req, res, next) => {
   // Create a case
+  const decodedToken = helper.tokenCheck(req, res);
+  if (decodedToken.affiliation !== 'faculty') {
+    return res.status(401).json({ error: 'you do not have rights to do this action' });
+  }
+
   const case1 = {
     title: req.body.title,
     hidden: req.body.hidden,
@@ -21,7 +25,7 @@ caseRouter.post('/', (req, res, next) => {
     .then((data) => {
       res.json(data);
     })
-    .catch((error) => next(error))
+    .catch((error) => next(error));
 });
 
 // Retrieve all cases
@@ -29,64 +33,42 @@ caseRouter.get('/', async (req, res, next) => {
   const { title } = req.query;
   const condition = title ? { title: { [Op.iLike]: `%${title}%` } } : null;
 
-  const data = await Case.findAll({ where: condition })
-  const user = {
-    user_name: req.headers.cn ? req.headers.cn : config.USER_NAME,
-    affiliation: req.headers.edupersonprimaryaffiliation ? req.headers.edupersonprimaryaffiliation : config.AFFILIATION,
-    studentid: req.headers.hypersonstudentid ? req.headers.hypersonstudentid : config.STUDENTID || '',
-    mail: req.headers.mail ? req.headers.mail : config.MAIL,
-  }
-
-  let token;
-
-  if (process.env.NODE_ENV !== 'test') {
-    const userFromDb = await User.findOrCreate({
-      where: {
-        user_name: user.user_name,
-        affiliation: user.affiliation,
-        studentid: user.studentid,
-        mail: user.mail
-      },
-      defaults: {
-        user_name: user.user_name,
-        affiliation: user.affiliation,
-        studentid: user.studentid,
-        mail: user.mail
-      }
-    })
-    const userForToken = {
-      username: userFromDb.user_name,
-      id: userFromDb.id,
-    }
-
-    token = jwt.sign(userForToken, process.env.SECRET)
-  }
+  const data = await Case.findAll({ where: condition });
+  const user = await helper.createUser(req);
 
   try {
     res
       .status(200)
-      .send({ token, name: user.user_name, admin: user.affiliation === 'faculty', data });
-  } catch(error) {
-    next(error)
+      .send({
+        user, data,
+      });
+  } catch (error) {
+    next(error);
   }
 });
 
 // Find a single case (by id)
 caseRouter.get('/:id', (req, res, next) => {
+  helper.tokenCheck(req, res);
   const { id } = req.params;
 
   Case.findByPk(id)
     .then((data) => {
       if (data === null) {
-        res.send(404).end()
+        res.send(404).end();
       }
       res.json(data);
     })
-    .catch((error) => next(error))
+    .catch((error) => next(error));
 });
 
-// Update a disease (by id)
+// Update a case (by id)
 caseRouter.put('/:id', (req, res, next) => {
+  const decodedToken = helper.tokenCheck(req, res);
+  if (decodedToken.affiliation !== 'faculty') {
+    return res.status(401).json({ error: 'you do not have rights to do this action' });
+  }
+
   const { id } = req.params;
 
   Case.update(req.body, {
@@ -99,11 +81,16 @@ caseRouter.put('/:id', (req, res, next) => {
         });
       }
     })
-    .catch((error) => next(error))
+    .catch((error) => next(error));
 });
 
 // Delete a case (by id)
 caseRouter.delete('/:id', (req, res, next) => {
+  const decodedToken = helper.tokenCheck(req, res);
+  if (decodedToken.affiliation !== 'faculty') {
+    return res.status(401).json({ error: 'you do not have rights to do this action' });
+  }
+
   const { id } = req.params;
 
   Case.destroy({
@@ -111,24 +98,29 @@ caseRouter.delete('/:id', (req, res, next) => {
   })
     .then((num) => {
       if (Number(num) === 1) {
-        res.status(204).end()
+        res.status(204).end();
       } else {
-        res.status(404).end()
+        res.status(404).end();
       }
     })
-    .catch((error) => next(error))
+    .catch((error) => next(error));
 });
 
 // Delete all cases
 caseRouter.delete('/', (req, res, next) => {
+  const decodedToken = helper.tokenCheck(req, res);
+  if (decodedToken.affiliation !== 'faculty') {
+    return res.status(401).json({ error: 'you do not have rights to do this action' });
+  }
+
   Case.destroy({
     where: {},
     truncate: false,
   })
-    .then((nums) => {
-      res.status(204).end()
+    .then(() => {
+      res.status(204).end();
     })
-    .catch((error) => next(error))
+    .catch((error) => next(error));
 });
 
 module.exports = caseRouter;
